tests/validation/requirements.txt
# Test validation dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
pyyaml>=6.0
